add betrayed status effect
