+++ ton
@@ -1,0 +1,252 @@
+{
+	"ranks": {
+		"stone": [0],
+		"iron": [1],
+		"gold": [2],
+		"emerald": [3],
+		"diamond": [4]
+	},
+
+	" - HERMITS - ": "Ordered by type and cost",
+
+	"BALANCED": "",
+	"vintagebeef_common": 0,
+	"cubfan135_common": 0,
+	"bdoubleo100_rare": 1,
+	"ethoslab_common": 1,
+	"hypnotizd_common": 1,
+	"rendog_common": 1,
+	"iskall85_common": 1,
+
+	"BUILDER": "",
+	"bdoubleo100_common": 0,
+	"zombiecleo_common": 0,
+	"grian_common": 0,
+	"keralis_common": 0,
+	"pearlescentmoon_common": 0,
+	"stressmonster101_common": 0,
+	"welsknight_common": 0,
+	"geminitay_common": 1,
+	"vintagebeef_rare": 1,
+	"falsesymmetry_rare": 1,
+	"goodtimeswithscar_rare": 1,
+	"rendog_rare": 2,
+
+	"EXPLORER": "",
+	"ijevin_common": 0,
+	"joehills_common": 0,
+	"xbcrafted_rare": 1,
+	"zedaphplays_rare": 2,
+	"vintagebeef_ultra_rare": 2,
+
+	"FARM": "",
+	"impulsesv_common": 0,
+	"xisumavoid_common": 0,
+	"iskall85_rare": 0,
+	"tangotek_rare": 1,
+	"docm77_rare": 2,
+	"joehills_rare": 3,
+
+	"MINER": "",
+	"tinfoilchef_common": 0,
+	"tinfoilchef_rare": 2,
+	"hypnotizd_rare": 3,
+	"tinfoilchef_ultra_rare": 2,
+
+	"PRANKSTER": "",
+	"grian_rare": 2,
+	"mumbojumbo_rare": 3,
+	"stressmonster101_rare": 3,
+
+	"PVP": "",
+	"falsesymmetry_common": 0,
+	"xbcrafted_common": 0,
+	"welsknight_rare": 2,
+	"zombiecleo_rare": 3,
+	"ethoslab_ultra_rare": 3,
+
+	"REDSTONE": "",
+	"docm77_common": 0,
+	"mumbojumbo_common": 0,
+	"tangotek_common": 0,
+	"zedaphplays_common": 0,
+	"xisumavoid_rare": 2,
+	"ethoslab_rare": 3,
+	"impulsesv_rare": 4,
+
+	"SPEEDRUNNER": "",
+	"cubfan135_rare": 1,
+	"ijevin_rare": 1,
+
+	"TERRAFORM": "",
+	"keralis_rare": 1,
+	"goodtimeswithscar_common": 1,
+	"geminitay_rare": 1,
+	"pearlescentmoon_rare": 2,
+
+	" - EFFECTS - ": "",
+
+	"composter": 0,
+	"curse_of_binding": 0,
+	"gold_armor": 0,
+	"instant_health": 0,
+	"iron_sword": 0,
+	"invisibility_potion": 0,
+	"knockback": 0,
+	"loyalty": 0,
+	"milk_bucket": 0,
+	"splash_potion_of_healing": 0,
+	"water_bucket": 0,
+
+	"wolf": 1,
+	"bow": 1,
+	"chorus_fruit": 1,
+	"crossbow": 1,
+	"curse_of_vanishing": 1,
+	"diamond_sword": 1,
+	"efficiency": 1,
+	"flint_&_steel": 1,
+	"fortune": 1,
+	"lead": 1,
+	"mending": 1,
+	"spyglass": 1,
+	"looting": 1,
+
+	"chest": 2,
+	"emerald": 2,
+	"fishing_rod": 2,
+	"golden_apple": 3,
+	"golden_axe": 2,
+	"instant_health_ii": 2,
+	"iron_armor": 2,
+	"shield": 2,
+	"tnt": 2,
+	"bed": 2,
+	"splash_potion_of_poison": 2,
+	"thorns": 2,
+
+	"diamond_armor": 3,
+	"lava_bucket": 3,
+	"netherite_sword": 3,
+	"totem": 3,
+
+	"clock": 4,
+	"netherite_armor": 4,
+
+	" - ITEMS - ": "",
+
+	"item_prankster_common": 0,
+	"item_miner_common": 0,
+	"item_redstone_common": 0,
+	"item_terraform_common": 0,
+	"item_farm_common": 0,
+	"item_speedrunner_common": 0,
+	"item_builder_common": 0,
+	"item_pvp_common": 0,
+	"item_balanced_common": 0,
+	"item_explorer_common": 0,
+
+	"item_balanced_rare": 2,
+	"item_builder_rare": 2,
+	"item_explorer_rare": 2,
+	"item_farm_rare": 2,
+	"item_miner_rare": 2,
+	"item_prankster_rare": 2,
+	"item_redstone_rare": 2,
+	"item_terraform_rare": 2,
+	"item_speedrunner_rare": 2,
+	"item_pvp_rare": 2,
+
+	" - ALTER EGOS HERMITS - ": "",
+
+	"beetlejhost_common": 1,
+	"eviljevin_common": 0,
+	"evilxisuma_rare": 4,
+	"helsknight_rare": 2,
+	"humancleo_rare": 2,
+	"goatfather_rare": 2,
+	"hotguy_rare": 1,
+	"jingler_rare": 1,
+	"llamadad_rare": 2,
+	"potatoboy_rare": 1,
+	"poultryman_common": 0,
+	"renbob_rare": 2,
+
+	" - ALTER EGOS EFFECTS - ": "",
+
+	"anvil": 0,
+	"command_block": 0,
+	"ender_pearl": 0,
+	"fire_charge": 0,
+	"piston": 0,
+
+	"bad_omen": 1,
+	"chainmail_armor": 1,
+	"egg": 1,
+	"splash_potion_of_healing_ii": 1,
+	"turtle_shell": 1,
+
+	"target_block": 3,
+	"armor_stand": 2,
+	"ladder": 2,
+	"lightning_rod": 2,
+	"potion_of_slowness": 2,
+	"potion_of_weakness": 2,
+	"string": 2,
+	"sweeping_edge": 2,
+	"trident": 2,
+
+	"thorns_ii": 3,
+
+	"thorns_iii": 4,
+
+	" - ADVENT OF TCG HERMITS - ": "",
+
+	"dungeontango_rare": 2,
+	"pythongb_rare": 3,
+	"orionsound_rare": 1,
+	"solidaritygaming_rare": 2,
+	"monkeyfarm_rare": 1,
+	"biffa2001_rare": 3,
+	"smajor1995_rare": 0,
+	"pixlriffs_rare": 1,
+	"bigbst4tz2_rare": 2,
+	"ldshadowlady_rare": 1,
+	"shubbleyt_rare": 1,
+	"smallishbeans_rare": 2,
+	"grianch_rare": 3,
+	"pharaoh_rare": 2,
+
+	" - ADVENT OF TCG EFFECTS - ": "",
+
+	"brewing_stand": 1,
+	"furnace": 1,
+	"cat": 1,
+	"berry_bush": 2,
+	"slimeball": 0,
+	"trapdoor": 2,
+
+	"dropper": 0,
+	"fletching_table": -1,
+	"brush": 0,
+	"glowstone": 2,
+	"lantern": 3,
+	"splash_potion_of_harming": 3,
+
+	" - ALTER EGOS II HERMITS - ": "",
+
+	"architectfalse_common": 0,
+	"boomerbdubs_rare": 1,
+	"dwarfimpulse_common": 0,
+	"fiveampearl_rare": 1,
+	"frenchkeralis_common": 0,
+	"horseheadhypno_common": 1,
+	"iskallman_rare": 1,
+	"originalxb_common": 1,
+	"overseer_rare": 0,
+	"princessgem_common": 1,
+	"shadee_common": 0,
+	"spookystress_common": 0,
+	"steampunktango_common": 0,
+	"wormman_common": 0
+}
